version 0.0.1 --> 0.1.0 :) Also small test cleanup
